… pandas/_libs/tslibs/np_datetime.pyi