```
In [1]: import pandas as pd

In [2]: dti = pd.date_range("2016-01-01", periods=10_000)

In [3]: td = pd.Timedelta(days=1)

In [4]: %timeit dti + td
108 µs ± 1.85 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)  # <- main
94.3 µs ± 3.27 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)  # <- PR
```